service: make indexer queue name consistent
